[new-signatures] Remove unnecessary TF2.0 errors
Summary:
To align with the existing behavior of TF1.0, this diff removes a lot of error categories from TF1.0. These are things like `TArgArity` errors, which will cause errors at check time anyways, so generating a signature verification error seems unnecessary.

We make this change by creating a new `errno`, `CheckError`, which represents things that will cause errors at check-time anyways. When errnos are converted to user-visible signature verification errors, `CheckError`s are filtered out.

With this change, TF2.0 can almost express all the verification errors it currently generates in terms of the existing `Signature_error.t` data structure. The only remaining `errno` that can't yet be converted to either a silent `CheckError` or to `SigError` using `Signature_error.t` is `UnsupportedKeyMirrorProp`, which is partially dealt with in D23737473.

Reviewed By: samwgoldman

Differential Revision: D23720634

fbshipit-source-id: 7d57aaa86ee0e6c6e2120b63042274f66b1e20a9