[new-signatures] Ability to use signature_errors from TF2.0
Summary:
This diff adds a new `errno` to `Type_sig.ml`, `SigError`. This error kind contains a `Signature_error.t`, the same kind of error that was used in TF1.0. Since `Signature_error.t` contains `locs`, this diff also changes that marking and packing code to operate over those values; fortunately this is made relatively simple by the pre-existence of a mapping function for `Signature_error.t`.

When an `errno` is converted to a printable error in `parsing_service_js.ml`, `SigErrors` are directly converted by just looking up the locs.

This diff doesn't use `SigError` anywhere, but see the next diff.

Reviewed By: samwgoldman

Differential Revision: D23326312

fbshipit-source-id: f81aeeb97cb32c185855734f6477a28a534a61dc