[new-signatures] Improve support for KeyMirror errors in both versions of TF
Summary:
This diff improves the correspondence between TF1.0 and TF2.0 for uses of `keyMirror`, so that inappropriate properties in calls to `keyMirror` in TF2.0 give the same useful error message as TF1.0.

There is unfortunately still divergence in the behavior in other cases, though, and this diff adds a test case to document that difference. In a program like
```
const ba = {a: 42};
export const b = keyMirror({...ba});
```
TF1.0 doesn't raise a signature verification error. However, TF2.0 does do so, pointing at `{...ba}`. It's not straightforward for TF2.0 to match 1.0 here, because TF2.0's handling of `keyMirror` works by building up literal object properties, which is a problem when having to evaluate an arbitrary expression. There may be a way to do so, but I'm out of my depth on the semantics and usage of `keyMirror`.

(^^^ outdated, see discussion)

Reviewed By: samwgoldman

Differential Revision: D23737473

fbshipit-source-id: cc634e8a198cb38900c152110c811fd2342727fd