Treat temporary objects with spreads like literal values
Summary:
The `$TEMPORARY$object<>` constructor is a temporary hack used by types-first to
make object literals look like annotations, in order to create a signature AST.

Before this diff, exporting a value `{...o}` would be converted to an annotation
`$TEMPORARY$object<{|...o|}>`. However, the result of a spread is not always
exact. If `o` is inexact, then the result of the spread should also be inexact.
This is how the `target = Value _` option in `SpreadType` behaves.

Reviewed By: panagosg7

Differential Revision: D23726477

fbshipit-source-id: 4bb258955f614a7003d7710fd8ab40647a5f59ff