Kill ObjFreezeT
Summary:
ObjFreezeT is a constraint that we emit when we see an AST for
`Object.freeze({...})`. That is, specifically when the `Object.freeze` function
is applied to a literal object expression.

The constraint worked by first creating an type for the object literal, then
passing that type as the lower bound to `ObjFreezeT`, which converts to object
to be frozen, and passes the frozen type to it's output variable.

This diff takes a different approach: instead, we pass the "frozen" flag as
context into the code which creates an object type for the literal expression.
This context is threaded into the code for spreads as well.

This diff also applies the same logic to the `$TEMPORARY$Object$freeze<>` type,
which is used internally as part of types-first. We create an annotation of this
type as part of a signature AST when we see an exported expression with the form
`Object.freeze({...})`.

Before this diff, the temporary type's handling would create a type from the
object annotation, then pass that type through `ObjFreezeT`. With this change,
we instead pass the "frozen" context down, while also treating an object
annotation more like an object literal.

Reviewed By: jbrown215

Differential Revision: D23726472

fbshipit-source-id: e3bb4d3d73ee3fc11a2ad1dcd117f304f8f56e87