Fix reasons when converting remote ref signatures into types
Summary:
Before this diff, the added test had a different error in new signatures mode.
Instead of the expected output, the error would print as:

```
Cannot instantiate Named import from module `./type_exports` because `T` [1] is not a polymorphic type.
```

It turns out that the complicated `RNamedImportedType` reason description, and
related descs, are only used for intermediate tvars and probably don't need to
exist at all.

Reviewed By: panagosg7

Differential Revision: D23827786

fbshipit-source-id: a1e3f585514f7e66a1239b67b8afcca32693ed84