Fix location of temporary types when converting signatures to types
Summary:
The logic in Type_annotation uses the location of the type argument as the
location of the resulting type when parsing $TEMPORARY$string annotations.
Before this change, Type_sig_merge used the location of the entire (pseudo) type
application.

This diff changes Type_sig_merge to match the behavior of types-first. If the
wrong type argument is provided, or the wrong number of arguments, we do create
an `any` type with the location of the entire type application in both cases,
although that should hopefully not be observable in errors.

Reviewed By: panagosg7

Differential Revision: D23829819

fbshipit-source-id: e2de07a24545dbf883895f12585d8e00b5f984b5