Add static name field to classes when converting sigs to types
Summary:
For classes, we add a static name field by default -- you don't need to write
`static name: string`. This mirrors the behavior of the runtime, where all
classes have an own static name property.

For interfaces, we should not do this, but we do for historical reasons. There
is a comment in Type_annotation that says we should not do it. I also added a
comment to Type_sig_merge.

Leaving the actual fixing for future work. For now, this diff just makes
types-first 2.0's behavior consistent.

Reviewed By: panagosg7

Differential Revision: D23827784

fbshipit-source-id: 85d0e7d470f17af1859cd0010caf4e7bc6f4554d