no assert-ground for types-first
Summary:
The assert-ground check was useful in classic mode, to ban exports where
type variables appeared in input positions (missing-annot). In types-first, this
check is largely subsumed in the signature-verification-failures. Also, because
of the way the "merge_component" code is being shared by the "merge" and
"check" phase, this check was run twice, which was completely redundant.

This diff removes this check in types-first mode completely.

In theory we should error in the exact same locations as before. However, due to
implementation discrepancies, there are some cases where we'll no longer raise
an error. That is, cases where we used to flag a missing-annot error, but we
won't raise any error now:

1. React component propTypes are skipped in signature verification
2. Some cases of implicit instantiation
3. Missing annotations due to reassignment. In types-first we don't allow
exported variables to be reassigned. This is a cannot-reassign-export error.
4. Instantiation of static members--since we currently allow static memebers
to carry generic types.
5. Unannotated functions returning functions are only flagged once in
types-first

Reviewed By: samwgoldman

Differential Revision: D23830520

fbshipit-source-id: dd2c59f687290900631eca40c9d1f7703dd30297