Track exported env in new signatures mode
Summary:
types-first does not track re-assignment of functions and classes. In order to
ensure that signatures are correct, we ban re-assignment of exported functions
and classes, making them effectively const.

This check is enabled by storing the locations of those declarations in the
file sig, which are then passed along to the context and used to find
re-assignments during check.

In vanilla types-first, we pass through all locations for all bindings in the
"env" data structure. In new signatures mode, we need to drive the locations of
exported declarations from the type sig data structures, which doesn't have a
direct analog for types-first's env.

We might not have an env, but we do have a list of "local defs," which includes
all classes and functions that are reachable from exports.

Reviewed By: panagosg7

Differential Revision: D23886339

fbshipit-source-id: 4edd9b422e1fa2ea2e264e8659599e1eb97feddc