Don't hoist block-level function declarations
Summary:
This diff changes the types-first signature builder to match the behavior of
statement.ml.

As noted in the comment, the "correct" behavior here is more complicated in
reality. In sloppy mode, a block-level function declaration will be hoisted. In
strict mode, things are more complicated. In ES5, it is an error to define a
block-level function declaration. In ES6+, the function declaration is local to
the block.

In statement.ml, the behavior matches ES6+ in strict mode. Depending on how the
source is actually compiled and what environment it runs in, this behavior may
not reflect the runtime behavior. Improving this situation is left for future
work.

For now, this diff only aims to make Flow internally consistent.

Reviewed By: panagosg7

Differential Revision: D23886340

fbshipit-source-id: d70eaa7f495f4bbdbdf57b11c623887b879a5e63