prune ignored directories if there are no negated ignores
Summary:
if there are no negated ignores, then if a parent directory matches an ignore then everything under it will also match.

in this case, we can prune the file search and avoid spending time walking a tree where everything will be ignored.

if there are negated ignores, we can't make this optimization. consider:

```
[ignores]
.*/foo/.*
!.*/bar/.*
```

we have to walk all of `foo/` in case we find a path like `foo/baz/bar/file.js`.

### Future work

if the negation is absolute, we could prune any ignored paths that are not a prefix of the negated ignore. for example:

```
[ignores]
.*/foo/.*
!/path/to/foo/bar/.*
```

when we visit `/path/to/foo/`, we see that it's ignored, but a prefix of the negation, so we need to look inside it. however, when we visit `/path/to/foo/baz/`, it is ignored and not a prefix of the negation, so we can prune that one.

this latter optimization is left for future work.

Reviewed By: samwgoldman

Differential Revision: D23877735

fbshipit-source-id: 643f1d115dca384165332412477e89a1cff07730