Change location for enum reasons
Summary:
Before this change, the reason for an enum was positioned at the entire
declaration. When an enum appears in an error message, the code reference to the
enum would point to the entire declaration. When printing errors with snippets,
the errors would be quite verbose.

This diff changes that location to point to only the name in the enum
declaration. Errors now only need to point to a location span on a single line
instead of multiple lines.

This makes errors a bit more compact, free of arguably superfluous detail. It
also makes enums more consistent with classes, which also point to the name
instead of the entire declaration.

Finally, this makes types-first 2.0 consistent, since types-first 2.0 was
already using the name location. It would have been possible to change the
behavior of Type_sig_merge instead, and I can still do that, but I prefer this
option and wanted to pursue it first.

Reviewed By: gkz

Differential Revision: D24033447

fbshipit-source-id: 4f574c2619a08827ee97b7a785989629dcfe5908