Fix unchecked node module vs. lib tests
Summary:
This test had a cmd specified in the .testconfig, which means the test does not
actually run properly in the test_saved_state or test_new_signatures test
targets.

The cmd was useful for this test because it works around the default
"no_flowlib" configuration in runtests.sh and also prevents the full path to the
lib/node.js file from being printed in the recorded errors.

Both of these issues can be worked around by instead specifying a lib.js file in
under the test's root, which is what this test does.

Reviewed By: mvitousek

Differential Revision: D24029226

fbshipit-source-id: 28f825480b6a60ac815a947d8f069bcf70bbaafe