Fix unchecked haste module vs. lib tests
Summary:
This test was supposed to be testing a behavior where a declared module in
a libdef will take precedence over an untyped haste module with the same name.

However, because the server was being run with --all, the haste module was
actually typed, and was being selected as the module provider.

This mistake seems to be original to the test itself, in D2615141 (https://github.com/facebook/flow/commit/189b109c422fc3fcf8bec97a9598c40ca02add01). Error
messages were bad enough then that it was difficult to tell, I guess.

Reviewed By: mvitousek

Differential Revision: D24029225

fbshipit-source-id: 7235d63674cc24bdf0e19e3a6fbc68ebcc776899