Fix legacy unchecked module handling
Summary:
The behavior of types-first 2.0 merge did not exactly match the behavior of
check when it comes to requiring an unchecked file. In this case, if there is a
libdef, the libdef "overrides" the unchecked file so we return the module
defined in the libdef instead of an untyped module.

The way that the fallback logic is implemented is a bit tricky. Requiring an
unchecked module emits a lookup constraint on the builtins object. We look up
the "internal module name" based on the module reference in the file. That is,
if we have `require("foo")`, then we look up `..$module__foo`. This will find
any declared modules from the libdef. If the lookup fails, it returns `any`.

In types-first 2.0, we were constructing the wrong internal module name. Instead
of `..$module__foo`, we looked for `..$module__/absolute/path/to/foo.js`. This
would naturally always fail and return any.

Reviewed By: mvitousek

Differential Revision: D24029735

fbshipit-source-id: 909d17440cf2aa874644c5f58ee74641e6bfdb67