[new-generics] Include generic ID in reasons for generate-test types
Summary:
We want to be able to detect when generics escape from their bounds (cf D21285146) but in the generate-tests world, how do we tell what types are generics and what scope they're in? This diff adds this information to the `RPolyTest` reason desc, which is the reason of all generic types created during generate-tests. The reason now contains the loc of the generic's definition site for use as a generic ID (ideally this would be an ALoc.id, but since the reason desc type is polymorphic over locs, I just store it as a ALoc.t and then cast it to ALoc.id when needed in a later diff).

This diff also adds a flag to both `RPolyTest`s and to the `typeparam` data structure to indicate whether the type parameter is `this` or not (since `this` is implemented in classes using generate-tests). This is purely to provide better error messages in D21285599.

Reviewed By: panagosg7

Differential Revision: D21285427

fbshipit-source-id: ea1da61d94789e65052f40f51cdd553257e95509