[new-generics] Tvars can have generic scopes
Summary:
This diff sets up the generic scope tracking machinery that is used in D21702070, which ultimately supports the generic escape analysis implemented in D21702072.

This diff provides a module, `Generic_scope`, that can be used to track what generics are in-scope at any given time during constraint generation and manages the context's scope map. This is analogous to the existing `Scope` module, except that 1. it focuses specifically on generics, which aren't handled by `Scope`, 2. it is constructed separately during the AST pass implemented in D21702072.

Reviewed By: panagosg7

Differential Revision: D21284706

fbshipit-source-id: 79a13db8d3aa05a10e931bc4444931e299b0762b