[new-generics] Make function/class types available to typed AST
Summary:
This diff supports an AST traversal in a later diff by making the general types for class and function identifiers available from the typed AST.

To show what's happening here, consider how a program like this interacts with the scope and the typed AST:
```
function f() { }
```
Here, `f` is stored in the environment as a value binding, with a general and a specific type. The specific type is a `FunT` generated from inspecting the function itself, and the general type is a tvar with the specific type as a lower bound.

Then, say we assign a value to `f`:
```
f = 42
```
Now, the general type for `f` will have `NumT` as a lower bound as well.

However, in the typed AST, currently the type given to the function ID `f` in the function declaration is the *specific* type of `f` at that point, the function type. This is a problem (and works differently from `var`s and the like) because we might want to traverse the AST, see that `f` is being introduced, and ask what types `f` could have as lower bounds. Currently, since we can only see the current specific type for `f`, the general type for `f` is unreachable (except as an upper bound) from the AST.

Work on generic escape prevention in D21702070 needs to see the general type for `f`, because it needs to be able to error if an out-of-scope `GenericT` becomes a lower bound for that general type. This diff enables that by making the general, not specific, type for function and class IDs become the typed AST entry for the ID in the class or function definition.

Class and function expressions still have the specific type as their overall expression type--it's only the ID AST node that is affected.

Reviewed By: samwgoldman, panagosg7

Differential Revision: D21702069

fbshipit-source-id: bf459852143c13e7234f9583b5018d03e57a9cb3