[new-generics] Create error messages for escaped generics
Summary:
Creates an error message for escaped generics, used in D21285719.

This feels like a somewhat challenging issue to explain to users so there are a few variations on how it's displayed. If a generic directly escapes to a variable like so:

```
var a;
function f<X>(x: X) {
  a = x;
}
```

then the error will be
```
Error --------------------------------------------------------------------------------- tests/generic_escape/zmsg.js:5:7

Cannot assign `x` to `a` because type variable `X` [1] cannot escape from the scope in which it was defined [2] (try
adding a type annotation to `a` [3]).

   tests/generic_escape/zmsg.js:5:7
   5|   a = x;
            ^

References:
   tests/generic_escape/zmsg.js:4:18
   4| function f<X>(x: X) {
                       ^ [1]
   tests/generic_escape/zmsg.js:4:12
   4| function f<X>(x: X) {
                 ^ [2]
   tests/generic_escape/zmsg.js:3:5
   3| var a;
          ^ [3]
```

However, if the generic escapes as a component of a larger type, it could be useful to point both at the larger type, and the generic's presence within that type: hence a program like
```
var a;
function f<X>(x: () => X) {
  a = x;
}
```

produces
```
Error --------------------------------------------------------------------------------- tests/generic_escape/zmsg.js:5:7

Cannot assign `x` to `a` because function type [1] contains type variable `X` [2] which cannot escape from the scope in
which it was defined [3] (try adding a type annotation to `a` [4]).

   tests/generic_escape/zmsg.js:5:7
   5|   a = x;
            ^

References:
   tests/generic_escape/zmsg.js:4:18
   4| function f<X>(x: () => X) {
                       ^^^^^^^ [1]
   tests/generic_escape/zmsg.js:4:24
   4| function f<X>(x: () => X) {
                             ^ [2]
   tests/generic_escape/zmsg.js:4:12
   4| function f<X>(x: () => X) {
                 ^ [3]
   tests/generic_escape/zmsg.js:3:5
   3| var a;
          ^ [4]
```

This diff also uses a different message when the generic that escaped is `this` rather than an explicit type parameter. Finally, the suggested fix may be absent if there isn't an annotation location recorded in the context.

Reviewed By: samwgoldman, panagosg7

Differential Revision: D21285599

fbshipit-source-id: 23b6432a937437e7c883c7a5d7b402d5270f4f5e