[new-generics] Create context module for use in tracking generic scopes
Summary:
For use in the generic escape framework, this diff creates a context-like module to track what tvars are in what generic scope. Currently this consists of a the `scope_map` from scope IDs to sets of ALoc.ids, which define what generics (represented by the ALoc.ids) are in any given scope; the `marked_ids`, a map from scope IDs to sets of tvars, which represent tvars that need to be checked for escaped generics and the scope that they're native to.

The `TvarFinder` visitor helps when adding tvars to the `marked_ids`. Given a type, it explores the type and collects all tvars it encounters *without* passing through any tvars. In D21702070, this is used when exploring the AST to determine which tvars need to be checked for escaped generics when given a node from the typed AST.

Reviewed By: panagosg7

Differential Revision: D21702071

fbshipit-source-id: a1806ea367781e76294d9a651694f92026a3da7a