[new-generics] Create AST pass to find writable tvars and assign them a generic scope
Summary:
(if this seems familiar, its because this was the description for D21285146. see below on how the implementation differs)

This diff uses the machinery from D21284706 to track what generics are in scope at each point during an AST traversal, and to add scope information to a subset of tvars by including them in the `marked_id` map of the generic context (D21702071). The actual checking that generics don't escape their scope happens in D21702072, but this diff does a lot of the important work to enable that diff, so I'll explain what's going on in this diff in a bit more detail.

The point of this work is to detect when generics escape from their scopes into tvars that can then be read from back inside the scope. The classic example of this is
```
var a = [];
function f<X>(x: X, should_escape: boolean): X {
  if (should_escape) {
    a[0] = x;
  }
  return a[0];
}

f<number>(42, true);
var b: string = f<string>("hello world", false); // but the result of this call is a number!
```
Currently, this results in [a (confusing) error](https://flow.org/try/#0G4QwTgBCELwQ2gXQNwCgBmBXAdgYwC4CWA9thOgDwAaAfABQAeAXBFQDQQDOAFsZgDYATAPoBTTrhAAHUSwBGxYv1EhsAShZUIAb1QQIhdBDo8+QsROmi1OvfqjwADIlgQGafQF87YUfkxgZCBOKKjeGBTYmAC2cqJg9AAsAEwc+GCY1migkHIsnOmE2ADmrpQFYEXF9ABE3KL8-MQQAO7EYEI1HOgg-JxZEAD0gxBymPgQ+PUQvpwCE8RGU4ScEJKNBqvQUbHxAIRAA) because generate-tests causes `a[0]` to have both `mixed` and `empty` as lower bounds, and so in the test of the function body where `X=empty`, we try to flow `mixed ~> empty` and get the error. When generate-tests is replaced by the new generic design, however, nothing will inherently cause this program to error, and the result would be the unsoundness shown above.

This diff adds the tracking information to say that the generic `X` is not in the scope of the element type tvar of `a`, so that D21702071 can raise an error.

There are two main points of subtlety here: what generics are in-scope when? And, which tvars need to have scopes in the first place?

### When is a generic in scope?

In the above example, it's obvious that `X` is not in scope for `a`. However, what about an example like this:
```
function g<X>(x: X) /* no return type annotation */ {
  return x;
}
```
It seems like the tvar representing the function's return type could have `X` be in scope--after all, it's fine to insert `: X` as an annotation there. However, there are benefits to saying that that tvar is actually out of scope, and demanding an annotation to be inserted instead, because it's still an avenue for the generic to escape, and most operations that one could perform using the escaped generic will result in [confusing error messages](https://flow.org/try/#0GYVwdgxgLglg9mABAcwDwA0B8AKAHgLkXQEpEBvAKEUQCcBTKEGpXAbgoF8KKA3AQxqI+hMCAC2AIzqCAvCmwAWAEzF2QA). This class of examples won't, as far as I've been able to analyze it, result in unsoundness that wouldn't be caught if we allowed the function return to have `X` in its scope, but by excluding it from the scope, we can get better-positioned error messages.

A similar argument applies for class properties.

### What generics have to have scopes?

In this design, the generic scoping system only applies to certain tvars, and the rest are treated as having all generics in their scope. This approach arises from both practical and semantic issues.

Probably the most critical issue is that this design for detecting escaped generics is a nonstarter in terms of performance if every tvar generated while checking a module needs to be examined for escaped generics--there's just too many. Additionally, for many tvars its not clear what scope they ought to be considered native to: certainly any tvars generated during constraint generation could be given scopes, but it's not clear what scope the connective tvars generated in `flow_js` belong to.

Ultimately though the point of this work is to catch the unsound cases of generics escaping their scope, and then being re-used within the scope:
```
                 tvar
                 ^  |
                 |  |
_________________|__V____
| scope of X            |
|_______________________|
```
and this can--or at least, *should*--only be able to happen with a subset of the tvars created during typechecking. For example, it ought to be fine for the tvar used as the `t_out` of a `CallT` to be unscoped--it won't be able to arbitrarily accumulate new upper and lower bounds, and even if it "switches sides" with via `typeof` it will still constrain writes in a way that prevents the above escape occurring without an error.

### Implementation

(this is the part that's different from D21285146)

This diff builds the `marked_ids` map in a generic context (D21702071) by performing a traversal over the typed AST. When the traversal encounters classes or functions with declared generics, those generics are installed in a new scope by the `ScopeHelper` internal helper module, and at every point during the traversal we can access the id for the current generic scope (with `Generic_scope.scope_id ()`).

When the AST traversal encounters certain nodes, such as unannotated variable declarations, it adds any tvars in the type of the node to the `marked_ids` map using `Generic_cx.mark_ids_of_type`. These tvars are then the tvars that will be checked for escaped generics (in a later diff).

In this diff, tvars are marked that appear in:
* places that can have a type annotation, but don't:
  * unannotated variable declarations
  * unannotated parameters
  * unannotated returns
  * unannotated class properties
* exception variables in `catch` clauses
* class and function names (because they can be written to like unannotated `var`s)
* object rest patterns (e.g. `var {a, b, ...c} = ...`; `c` needs to be marked)

Reviewed By: samwgoldman

Differential Revision: D21702070

fbshipit-source-id: 574ed95c404be667b22ca4301548060ea991f0fc