[new-generics] Track missing annotation locations and explore types for tvars to provide better escaped generics errors
Summary:
This diff improves the errors generated by the escape analysis of D21702072 by exploring types to find the tvars that are most likely to actually be responsible for generics escaping, and by providing suggested fixes.

### More precise error messages

We detect generic escapes by looking into the lower bounds of scoped tvars, as detected by the `Scoped_tvar_finder`. Originally, when we found an out-of-scope generic in a lower bound, we would raise an error pointing at the entire type that contained the escaped generic. So, for example, in this program:
```
var x = [];
function f<T>(t: T) {
  x[0] = t;
}
```
The error here is that `T` escapes into the elements tvar for the array `x`. The AST traversal marks `x`'s tvar as the one that needs to be inspected for escaped generics, and then the escape analysis explores `x`'s type, reaches the array elements tvar, and eventually finds `T`, raising an error. However, because the analysis started with `x`, the error message we get is
```
Cannot assign array literal to x because empty array literal [1] contains type variable T [2] which cannot escape from
the scope in which it was defined [3]. [escaped-generic]
```

This doesn't point to the actual cause of the escape, but rather to the introduction of the tvar into which a generic escaped, so it's not very helpful.

This diff changes the escape analysis to use the use op and reason closest to the site of the actual escape. Instead of using the reason and use ops of the tvar bounds that serve as an "entry point" to the escape analysis, now every time the search encounters a tvar and explores its lower bounds, it will keep the use op and the reason of the bound to use when reporting any escapes found within (as long as the use op is not `UnknownUse`--we want to make sure we report a useful use op if there's any to be found). The result is a more useful error message:

```
Cannot assign `t` to `x[0]` because type variable T [1] cannot escape from the scope in which it was defined [2].
[escaped-generic]
```

The reasons and use op for the error message now point to exactly the site of the generic escaping.

### Suggested fixes

This kind of error can be hard to fix, since it arises due to how Flow performs inference, which users shouldn't need to have a deep understanding of. To try and alleviate this, this diff tries--where possible--to indicate a position that should be annotated to fix the error (because the annotation would constrain writes and prevent the generic from becoming a lower bound to a tvar without an error being raised).

We can suggest annotations by recording the reasons for unannotated tvars when using the AST pass to find tvars that have to be scope checked. In cases where an annotation could be provided, but isn't, we record the reason of the missing annotation in the `blame_reasons`, a map from tvars to reasons (stored in the generic context).

Then when we search for escaped generics in `generic_escape.ml`, whenever we explore a tvar's lower bounds for escaped generics, we check if the tvar has a reason in the `blame_reasons` map. If so, we can report that an annotation there would prevent the escape.

As with the error reasons and use ops above, errors will report the closest annotation site to the actual cause of the escape. In some cases the annotation site is farther away than the use op: because there's no annotation to put on the elements tvar of an empty array literal, the best we can do for the above error is to recommend that the declaration site be annotated. The final error we get, then, is:

```
Cannot assign `t` to `x[0]` because type variable `T` [1] cannot escape from the scope in which it was defined [2]
(try adding a type annotation to `x` [3]). [escaped-generic]

      x[0] = t;
             ^

References:
   function f<T>(t: T) {
                    ^ [1]

   function f<T>(t: T) {
              ^ [2]

   var x = [];
       ^ [3]
```

Reviewed By: panagosg7

Differential Revision: D21702076

fbshipit-source-id: 741429d10233eaf9fa885af08b22fc13bdb81b74