Tweak declared var/function reasons
Summary:
This diff makes the reason for a declared var/function the same as it would be
for a non-declared version. A benefit of this is that types-first's effects are
less observable. Since types-first converts non-declare syntax to declare
syntax, the types that come from signature dependencies have observable
differences in error output.

This also makes types-first 2.0 consistent with types-first, since types-first
2.0 does not perform the conversion to declare syntax (or any AST-to-AST
conversion).

Reviewed By: panagosg7

Differential Revision: D24057449

fbshipit-source-id: c3b5afa8933e7bf9db50f2255d6b6b2207976995