[new-generics] Add the GenericT type and its baseline logic
Summary:
This diff implements the `GenericT` type and its basic logic, which will be used to replace the behavior of generate-tests in checking generics. At this point `GenericT`s are unused-they're first tested in D22436292.

There's going to be lots and lots of special cases and improvements on this design in later diffs, but the basic idea of a `GenericT` is captured here: a `GenericT` has an (upper) bound type and an id (as well as a reason and, for error message purposes only, a string name). The id is currently an ALoc.id, but will be replaced by a bespoke structure later.

`GenericT`s represent ranges of types, from `empty` to their upper bound. We should behave like the "actual" type of the generic is some unknown point in the space of types in that range, and ban any behavior that could be unsound at any point in that space. In this sense the `GenericT` is actually very similar to an `OpaqueT`, and indeed later on I plan to merge these behaviors.

When a `GenericT` is an upper bound to a constraint, it acts like `empty`, and when it's a lower bound, it acts like its bound. In practice this means nothing should be able to flow into a `GenericT`, which is appropriate because it could be any subtype of its upper bound, including `empty`. On the other hand, a `GenericT` can flow into positions that expect some supertype of its upper bound.

When one `GenericT` flows into another and the ids match, we flow the lower type's bound into the upper type's bound. This is needed because in later diffs we can perform operations on the bounds of `GenericT`s (e.g. refinement), and even if the matching ID proves that the two types come from the same source (e.g. same type parameter), we still need to check that the ranges of types that they can describe are in a subset relation: the upper bound `GenericT` represents a larger set of types that includes the range of the lower bound.

Reviewed By: panagosg7

Differential Revision: D22436118

fbshipit-source-id: e32f1f52934e8ec46c7bed8db212dfb8e139ed48