[new-generics] Test harness can switch the generate-tests flag
Summary: The new generics work should eventually behave similar or identically to the existing generate-tests behavior, so it's useful to conveniently be able to run it against the existing test suite without having to rebuild Flow. This diff adds a `-g` flag to run runtests.sh with the new generics behavior (can be overridden by .flowconfig).

Reviewed By: panagosg7

Differential Revision: D22436116

fbshipit-source-id: 16db8c8315590f6d5524d58e95827ee059e2221a