Add check_should_exit to the end of with_no_cancellations
Summary:
If you want a group of operations that can touch the shared heaps to be atomic, you can wrap them with `with_no_cancellations` to prevent cancellations from taking place during that operation. Unfortunately, if many of our shared heap operations are wrapped in this utility, there's the potential for cancellations to be delayed for a long time, with potential negative performance ramifications.

So, let's check if we need to cancel after each block of atomic operations.

Note that I had to rearrange the functions a bit so that the version of `check_should_exit` that exits cleanly is in scope.

Reviewed By: samwgoldman

Differential Revision: D24095226

fbshipit-source-id: 59e340023d8ffe324ef1bb2319cfda2b24eb6b76